Restoring the test suite with updates for Ruby 2.7.z releases and by introducing WebMock; Updating the README and adding the license
