+++ lib/tasks/oai.rake
@@ -4,6 +4,7 @@ require 'orangetheses'
 
 namespace :oai do
   desc 'Index Published Item using the OAI-PMH'
+  # rubocop:disable Rails/RakeEnvironment
   task :index_record, [:identifier] do |_task, args|
     harvester = Orangetheses::Harvester.new
     indexer = Orangetheses::Indexer.new
@@ -11,4 +12,5 @@ namespace :oai do
     identifier = args[:identifier]
     harvester.index_item(indexer: indexer, identifier: identifier)
   end
+  # rubocop:enable Rails/RakeEnvironment
 end
